// src/app/components/binance/interfaces/binance-coin-flexible.interface.ts
export interface BinanceCoinFlexibleInterface {
  avgAnnualInterestRate: string;
  tierAnnualInterestRate?: any;
  minPurchaseAmount: string;
  purchasedAmount: string;
  upLimitPerUser: string;
  canPurchase: boolean;
  canRedeem: boolean;
  productId: string;
  featured: boolean;
  upLimit: string;
  status: string;
  asset: string;
}