// src/app/components/binance/interfaces/binance-coin-details.interface.ts
export interface BinanceCoinDetailsInterface {
  withdrawAllEnable: boolean;
  depositAllEnable: boolean;
  isLegalMoney: boolean;
  withdrawing: string;
  networkList: any[];
  trading: boolean;
  ipoable: string;
  storage: string;
  freeze: string;
  ipoing: string;
  locked: string;
  coin: string;
  free: string;
  name: string;
}